@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 /*---------------------------------------------------------------------------------------------
@@ -23,12 +23,9 @@ let package = Package(
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
         .package(name: "itwin-mobile-native", url: "https://github.com/iTwin/mobile-native-ios", .exact("3.2.14")),
-        .package(url: "https://github.com/mxcl/PromiseKit", from: "6.15.3"),
-        .package(name: "PMKFoundation", url: "https://github.com/PromiseKit/Foundation.git", from: "3.0.0"),
-        // The following is a fork of CoreLocation that changes the iOS platform to v9
-        .package(name: "PMKCoreLocation", url: "https://github.com/fallingspirit/CoreLocation", from: "3.1.2"),
+        .package(url: "https://github.com/AsyncSwift/AsyncLocationKit.git", .upToNextMajor(from: "1.0.0")),
         .package(name: "Reachability", url: "https://github.com/ashleymills/Reachability.swift", from: "5.1.0"),
-        .package(name: "AppAuth", url: "https://github.com/openid/AppAuth-iOS.git", .upToNextMajor(from: "1.4.0")),
+        .package(name: "AppAuth", url: "https://github.com/openid/AppAuth-iOS.git", .upToNextMajor(from: "1.6.0")),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -37,17 +34,13 @@ let package = Package(
             name: "ITwinMobile",
             dependencies: [
                 .product(name: "IModelJsNative", package: "itwin-mobile-native"),
-                .product(name: "PromiseKit", package: "PromiseKit"),
-                .product(name: "PMKFoundation", package: "PMKFoundation"),
-                .product(name: "PMKCoreLocation", package: "PMKCoreLocation"),
+                .product(name: "AsyncLocationKit", package: "AsyncLocationKit"),
                 .product(name: "Reachability", package: "Reachability"),
                 .product(name: "AppAuth", package: "AppAuth"),
-            ],
-            resources: [
-                .process("ITwinMobile.docc")
             ]),
 //        .testTarget(
 //            name: "ITwinMobileTests",
 //            dependencies: ["ITwinMobile"]),
-    ]
+    ],
+    swiftLanguageVersions: [.version("5.5")]
 )
